+++ src/main/kotlin/io/github/durun/nitron/core/parser/antlr/AstBuildVisitor.kt
@@ -1,4 +1,4 @@
-package io.github.durun.nitron.core.parser
+package io.github.durun.nitron.core.parser.antlr
 
 import io.github.durun.nitron.core.antlr4util.children
 import io.github.durun.nitron.core.ast.node.AstNode
