+++ src/test/kotlin/io/github/durun/nitron/core/ast/processors/AstNormalizerTest.kt
@@ -3,8 +3,8 @@ package io.github.durun.nitron.core.ast.processors
 import io.github.durun.nitron.core.ast.path.AstPath
 import io.github.durun.nitron.core.ast.type.nodeTypePoolOf
 import io.github.durun.nitron.core.config.loader.LangConfigLoader
-import io.github.durun.nitron.core.parser.AstBuildVisitor
 import io.github.durun.nitron.core.parser.ParserStore
+import io.github.durun.nitron.core.parser.antlr.AstBuildVisitor
 import io.kotest.core.spec.style.FreeSpec
 import io.kotest.matchers.shouldBe
 import java.nio.file.Paths
