+++ src/test/kotlin/io/github/durun/nitron/core/parser/ParserStore.kt
@@ -1,6 +1,7 @@
 package io.github.durun.nitron.core.parser
 
 import io.github.durun.nitron.core.config.GrammarConfig
+import io.github.durun.nitron.core.parser.antlr.GenericParser
 import io.kotest.mpp.log
 import java.nio.file.Path
