+++ src/test/kotlin/io/github/durun/nitron/tester/ParserTester.kt
@@ -1,8 +1,8 @@
 package io.github.durun.nitron.tester
 
 import io.github.durun.nitron.core.ast.node.AstNode
-import io.github.durun.nitron.core.parser.AstBuildVisitor
 import io.github.durun.nitron.core.parser.ParserStore
+import io.github.durun.nitron.core.parser.antlr.AstBuildVisitor
 import org.snt.inmemantlr.exceptions.ParsingException
 import java.nio.file.Path
 import kotlin.io.path.ExperimentalPathApi
